# poslappapi/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /zonas:
    x-swagger-router-controller: zonas.controller
    get:
      description: Mostrar las zonas en lista (get the game system list)
      operationId: getZonas
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetZonasListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Agrega una nueva zona a la lista (add a new game system to the list)
      operationId: createZonas
      parameters:
        - in: body
          name: Zonas (gamesystem)
          description: La zona a crear (The Game System to create.)
          schema:
            $ref: "#/definitions/Zonas"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetZonasResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
  /zonas/{id}:
    x-swagger-router-controller: zonas.controller
    get:
      description: Dame una zona por id (get a game system)
      operationId: getZonasByPk
      parameters:
        - name: id
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetZonasResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Actualizar la zona por id (update a game system)
      operationId: updateZonas
      # define the parameters
      parameters:
        - name: id
          description: Zona Id (game system id)
          type: integer
          in: path
          required: true
        - in: body
          name: Zona (gamesystem)
          description: La zona a actualizar (The Game System to update.)
          schema:
            $ref: "#/definitions/Zonas"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetZonasResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Borrar una zona por id (delete a game system)
      operationId: deleteZonas
      # define the parameters
      parameters:
        - name: id
          description: Zona id (game system id)
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Zonas:
    type: object
    properties:
      id_zona:
        type: integer
        description: Id de la zona (Id of the Game System)
      nombreZona:
        type: string
        description: Nombre de la zona (Name of the Game System)
      descripcionZona:
        type: string
        description: Descricion de la zona (Description of the zone)
    required:
      - id_zona
      - nombreZona
      - descripcionZona
  GetZonasResponse:
    type: object
    properties:
      id_zona:
        type: integer
        description: Id de la zona (identifier)
      nombreZona:
        type: string
        description: Nombre de la zona (Name of the Game System)
      descripcionZona:
        type: string
        description: Descripcion de la zona (Description of the Game System)
    required:
      - id_zona
      - nombreZona
      - descripcionZona
  GetZonasListResponse:
    required:
      - gamesystems
    properties:
      gamesystems:
        type: array
        items:
          $ref: "#/definitions/GetZonasResponse"
  GeneralResponse:
    type: object
    properties:
      success:
        type: integer
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string